wext: Suppress kernel generated disassoc event
wpa_supplicant already generates the event internally when
disassociating. In the case of a new association, such an event may also
be generated by the kernel for the previous connection and needs to be
ignored.

Without this fix the next commit that optimizes the scanning behavior on
connection failures would trigger an unexpected failure in the
wext_pmska_cache hwsim test.

Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>