WNM: Extend the discovery of the 6 GHz candidates
Candidates on 6 GHz channels might be collocated ones, and thus, in
order to discover them need to include the frequencies on the 2.4/5 GHz
bands.

Extend the WNM scan frequency channels to include the 2.4/5 GHz channels
in case the WNM neighbor list included candidates on the 6 GHz band.
Since such a scan can be long, optimize the case of a single channel and
instead force passive scanning.

In addition, in case the candidate validity interval is too short,
ignore the validity interval, and set it to 5 seconds, as we are
increasing the discovery time so results might not be valid once we get
them.

Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-by: Andrei Otcheretianski <andrei.otcheretianski@intel.com>