supplicant: fix logic filtering multi-phy freqs
For multiphy radio, stations would scan all channels on all
bands, even when there was already an associated station on
one of the bands. This is due to the fact that the logic for
filtering bailed early when testing a frequencies availability
in a supported band. Fix is to fall through filter logic rather
than exit early.

Signed-off-by: Dylan Eskew <dylan.eskew@candelatech.com>