WPS ER: Clear HTTP socket on NOTIFY processing errors
Since we process only a single request from the socket, there is no need
to maintain the HTTP socket when we detect an error. That will just
delay closing of the socket until WPS ER is stopped. Instead, close the
socket immediately on detecting the error to match the behavior for
unsupported HTTP request types and any cases where we send a response.

Signed-off-by: Jouni Malinen <jouni.malinen@oss.qualcomm.com>