tests: Permit autogo_scan to work with dedicated P2P device removal
The next commit will end up removing the dedicated P2P device when
p2p_disabled is set. The test will trigger this when
support_p2p_device=1 is used and that would lead to a test failure as
the device is not configured correctly and may have changed its address.

Fix this by repeating the setup after the dedicated P2P device has been
recreated.

Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
Reviewed-by: Andrei Otcheretianski <andrei.otcheretianski@intel.com>