P2P: Add/remove dedicated P2P Device when p2p_disabled is toggled
Having a dedicated P2P Device interface can interfere with the operation
of other interface types (in particular NAN). As such, it is useful to
be able to toggle at runtime whether the dedicated P2P Device interface
exists.

Add the logic to add or remove the dependent dedicated P2P Device
interface when the p2p_disabled option is toggled on its parent
interface.

Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
Reviewed-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-by: Andrei Otcheretianski <andrei.otcheretianski@intel.com>