Include the mandatory part of MCS of HE Capabilities in struct
struct ieee80211_he_capabilities included the first four octets of the
Supported HE-MCS And NSS Set field in the optional[] array. However,
those four octets are actually required to be present. Add them as
mandatory fields into the struct to match the minimum set of fields and
to make them somewhat more convenient for parsing purposes.

Signed-off-by: sunilravi <sunilravi@google.com>
Signed-off-by: Henry Yen <henryyen@google.com>