Replace configuration int lists with int_arrays
This cleans up implementation by getting rid of very similar
construction of a list of int values. hostapd used to terminate the
lists with -1 while int_array were terminated with 0. There are no cases
where 0 is needed to be included, so all these can be converted into the
existing int_array design.

Signed-off-by: Jouni Malinen <jouni.malinen@oss.qualcomm.com>