nl80211: Only do station mode fallback for -EOPNOTSUPP
If the AP interface does not support scanning (or if
NL80211_SCAN_FLAG_AP is not set), then mac80211 will report -EOPNOTSUPP.
Only in this case it makes sense to try switching the interface type. So
change the restriction and drop the now unreachable code. Note that in
the case of MLD we will still never try to do an interface switch as
there is a check for valid_links later on.

This replaces the earlier changes from commit 666d695bbd04 ("nl80211:
Fix simultaneous scanning failure in case of MLO") with a simpler design
for avoiding the undesired change to STA mode in the AP MLD case.

Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>