nl80211: Support scanning as AP when not beaconing
If the kernel indicated support for scanning on AP interface, set
NL80211_SCAN_FLAG_AP in the scan request when the interface is not yet
beaconing. In that case, it is possible to do a scan without interfering
with normal AP operation so it is safe to set the flag.

This is important because otherwise it may not be possible to do the
appropriate scans required to bring up an interface that is part of an
MLD.

Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-by: Andrei Otcheretianski <andrei.otcheretianski@intel.com>