PASN: Fix testing code parsing for PASN_DRIVER
Incorrect indentation level hid the issue with the peer pointer not
being verified correctly. Fix the indentation level to make it clear
that peer might be NULL here and reject the cases that would have
resulted in dereferencting a NULL pointer. This code is included only
with CONFIG_TESTING_OPTIONS.

Fixes: ba7d967da46d ("PASN: Testing support for PASN with user-specified parameters")
Signed-off-by: Jouni Malinen <j@w1.fi>