USD: Support suspending USD operations
Based on the USD specification, the device should always be either on
the default channel or one of the configured channels. However, to
allow operation of other interfaces, configure the NAN Discovery
Engine (DE) to relax the USD availability requirements:

- Allow to configure N-Min and N-Max to different values than the
  default ones, to increase availability for the other interfaces
  (by configuring lower values).
- Allow to configure a suspend cycle, where after a certain amount of
  time that USD is active, it would be suspended for a certain amount
  of time.

While this might reduce the availability of USD, it allows the
operation of other interfaces that share the same radio.

Signed-off-by: Ilan Peer <ilan.peer@intel.com>