NAN USD: Allow services to disable listen operation during SDF exchange
The new stop-listen commands (through wpa_supplicant control interface
or D-Bus) can be used by an upper layer service implementation to stop a
listen operation (i.e., wait for a response SDF from the peer) in the
driver to free up the radio for other parallel operations at points in
the SDF exchange when the service knows the peer is not expected to send
any further SDFs before an SDF is sent to it.

This can be used, e.g., to optimize radio use during provisioning
exchange where the SDF exchange is stopped in the middle during the
connection attempt and the result of that connection attempt is
reporting later within that same SDF exchange.

Signed-off-by: Jouni Malinen <jouni.malinen@oss.qualcomm.com>