wpa_debug: Use separate buffer for path and improve error checking
Using the same buffer for output and input could already result in
an overlapping source and destination in snprintf. This was working
fine, however, with the patch to also find tracefs, we continue to parse
the buffer. In that case, the continued parsing can corrupt the found
path causing an error.

Fix this problem and reshuffle the code a bit to make it a bit more
clear and improve the condition to skip lines that could not be parsed
properly.

Fixes: 0a76c7ed64de ("wpa_debug: Prefer tracefs over debugfs")
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>