utils: Keep the last NOTE message as context for backtraces
When storing a backtrace, keep the string that was last used in a NOTE
message as context information. Keep all past strings in a list so that
they will remain available.

With this, we are able to show in leak reports which test did the
allocation. This is useful to more quickly solve issues as it can be
hard to find the triggering test otherwise.

Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>