MSCS: Reject MSCS ADD request when the MSCS session already is active
Add a check to block sending an MSCS ADD request if an MSCS session is
already active as this may lead to termination of the existing MSCS
session. This prevents termination of an existing MSCS session and
ensures robust handling of MSCS requests.

Also log a clear message indicating the reason for rejection.

Signed-off-by: Shubham Joshi <shujosh@qti.qualcomm.com>