DPP: Fix error path processing in decryption of E-id
If any of the operations on a successfully allocated e_id fails
dpp_decrypt_e_id() freed e_id, but forgot to clear the pointer that gets
returned to the caller. This could result in at least a theoretical use
of freed memory. Fix this by clearing the pointer to return an
appropriate indication of the failed operation.

Signed-off-by: Burvin Shi <burvin@amazon.com>