SAE: Use defines for authentication transaction sequence number
Use defined values instead of hardcoded 1 and 2 for the authentication
transaction sequence number of the SAE commit and confirm messages. This
makes the implementation easier to read and has a bit more context on
uses of this field for SAE-specific vs. general (i.e., applied to all
authentication algorithms) cases.

Signed-off-by: Jouni Malinen <jouni.malinen@oss.qualcomm.com>