VLAN: Group state machine data creation for partner AP MLD links
In EAPOL-Key msg 3/4 and group handshake message 1, the AP needs to
populate group keys from the appropriate VLAN group state machine data
(wpa_group) for each of the setup links. In the existing dynamic VLAN
implementation, when a multi-link STA associates on a particular VLAN,
VLAN group is created for the association link, but the same is not
created for partner links. As a result, in the absence of VLAN group
state machine data in partner authenticator object, at the time of
fetching group keys for the non-association links, in
wpa_auth_ml_get_key_info(), the helper function
wpa_select_vlan_wpa_group() fetches group keys on non-association links
from the default VLAN group.

Address this by allowing the ap_sta_set_vlan() and ap_sta_bind_vlan()
functionality to repeat for each of the setup links, i.e., as part of
ap_sta_set_vlan(), perform VLAN assignment to STA object in each of the
setup links and also allow VLAN group state machine data creation in the
partner authenticators as part of ap_sta_bind_vlan().

Fixes: dd65e53c9476 ("VLAN: Use VLAN group keys for EAPOL frames and FT reassoc for MLO")
Co-developed-by: Adil Saeed Musthafa <adilm@qti.qualcomm.com>
Signed-off-by: Adil Saeed Musthafa <adilm@qti.qualcomm.com>
Signed-off-by: Sai Pratyusha Magam <smagam@qti.qualcomm.com>