tests: Fix nan_usd_provisioning
This test case had swapped the subscribe and publish IDs for some of the
operations and expected values and that resulted in this failing
whenever dev[0] and dev[1] did not have the exact same current ID value
in their NAN state. For example, this failed in this test case sequence:
nan_usd_publish nan_usd_provisioning

Fix this by using the correct handle IDs for the service on each device.

Signed-off-by: Jouni Malinen <jouni.malinen@oss.qualcomm.com>