PASN: Calculate Auth1 hash after PTK derivation
It is more convenient to wait with the Auth1 hash calculation to avoid
having to figure out which hash algorithm to use before the PTK is
derived. Auth1 hash is defined to use the same hash algorithm as the one
that was used during PTK derivation. This requires a bit more memory,
but that is justifiable with the simplified implementation.

Signed-off-by: Jouni Malinen <jouni.malinen@oss.qualcomm.com>