SME: Do not send RSNXE for WPA 1 connections
The element is not applicable to WPA 1 and it will not be included in
the KDE later on. However, hostapd checks that the KDE and the
association request match and deauthenticates the station if it sent the
RSNXE in the association request.

So omit sending it if the protocol is WPA_PROTO_WPA.

Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>