nl80211: Support NAN Device interface
Add support for getting NAN Device interface capabilities from the
kernel and support adding a NAN Device interface.

As the NAN device interface is created with NL80211_ATTR_SOCKET_OWNER,
meaning that the NAN events would be sent on the socket that was used
for creating the interface, use a dedicated socket for the NAN Device
interface creation, to avoid races between kernel asynchronous events,
and kernel responses for commands sent from wpa_supplicant.

Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Signed-off-by: Andrei Otcheretianski <andrei.otcheretianski@intel.com>