NAN: Handle cluster join and next DW events
To support synchronized NAN Discovery Engine (DE) in wpa_supplicant, the
driver will report two new events:

NAN_CLUSTER_JOIN: This event is sent whenever a new NAN cluster is
started or joined. This event carries the new cluster id, which will be
used to construct SDFs.

NAN_NEXT_DW: A notification about an upcoming Discovery Window (DW).
This event is optional and may be turned on and off. It is used to
trigger multicast SDF transmissions during the upcoming DW.

Define these events and add the functions needed for delivering thenm to
nan_supplicant.c

Signed-off-by: Andrei Otcheretianski <andrei.otcheretianski@intel.com>